refactor(src): group into storage/ and stac/ module folders
