+++ src/lib.rs
@@ -1,13 +1,8 @@
-pub mod catalog;
 pub mod error;
-pub mod parquet_io;
-pub mod pmtiles;
-pub mod registry;
-pub mod s3;
-pub mod schema;
-pub mod theme;
-
-pub use error::{Error, Result, ResultExt};
+pub mod stac;
+pub mod storage;
 
 #[cfg(feature = "python")]
 mod python;
+
+pub use error::{Error, Result, ResultExt};
