+++ tests/bucket.rs
@@ -1,6 +1,6 @@
 //! Pure unit tests for `Bucket::from_url` — no network, no I/O.
 
-use overture_stac::s3::Bucket;
+use overture_stac::storage::Bucket;
 use overture_stac::Error;
 
 fn err_msg<T>(r: Result<T, Error>) -> String {
